Started on the shift_between_days neighbor
